Backend_Test/requirements.txt
pytest==8.1.1
requests==2.31.0
google-generativeai==0.4.1
python-dotenv==1.0.1
pydantic==2.6.4
